Move loadExtensions to own module
