+++ lib/utils/index.js
@@ -1,9 +1,10 @@
+'use strict';
 
-var crypto = require('crypto')
-  , _ = require('lodash')
-  , extend = require('extend')
+var crypto = require('crypto');
+var _ = require('lodash');
+var extend = require('extend');
 
-module.exports ={ 
+module.exports = { 
     gravatar: gravatar
   , sanitizeProject: sanitizeProject
   , sanitizeBranch: sanitizeBranch
